refactor: isolate Responses API handling
