+++ gen.pollinations.ai/src/text/handler.ts
@@ -6,39 +6,27 @@ import {
     FALLBACK_TARGET_HEADER,
     MODEL_USED_HEADER,
     openaiUsageToUsage,
-    responsesUsageToUsage,
 } from "@shared/registry/usage-headers.ts";
-import type {
-    CreateChatCompletionRequest,
-    CreateResponseRequest,
-} from "@shared/schemas/openai.ts";
+import type { CreateChatCompletionRequest } from "@shared/schemas/openai.ts";
 import type { Context } from "hono";
-import type { ContentfulStatusCode } from "hono/utils/http-status";
 import type { Env } from "@/env.ts";
 import {
     attachFallbackTarget,
     type FallbackCandidate,
     fallbackCandidates,
-    formatFallbackTarget,
     withModelFallback,
 } from "../fallback.ts";
 import { fixWavHeader } from "../routes/audio.js";
 import type { GenerateTextRequestQueryParams } from "../schemas/text.ts";
 import { enforceModelRateLimit } from "../utils/model-rate-limit.ts";
 import { communityEndpointGatewayContext } from "./communityEndpoint.ts";
+import { syncTextEnvironment } from "./environment.js";
+import { throwTextError } from "./errors.js";
 import { generateTextPortkey } from "./generateTextPortkey.js";
 import {
     getChatRequestData,
     getSimpleTextRequestData,
 } from "./requestUtils.js";
-import {
-    callDirectResponses,
-    type DirectResponsesTarget,
-    getResponseUsage,
-    resolveDirectResponsesTarget,
-    responsesInvalidRequest,
-    validateDirectResponsesRequest,
-} from "./responses.js";
 import type {
     ChatCompletion,
     RequestData,
@@ -47,38 +35,6 @@ import type {
 } from "./types.js";
 
 type TextContext = Context<Env>;
-
-const TEXT_ENV_KEYS = [
-    "AI_GATEWAY_API_KEY",
-    "AWS_ACCESS_KEY_ID",
-    "AWS_REGION",
-    "AWS_SECRET_ACCESS_KEY",
-    "AZURE_MYCELI_PROD_API_KEY",
-    "AZURE_MYCELI_PROD_SWEDEN_API_KEY",
-    "DASHSCOPE_API_KEY",
-    "DEEPINFRA_API_KEY",
-    "FIREWORKS_NEO_API_KEY",
-    "GOOGLE_CLIENT_EMAIL",
-    "GOOGLE_PRIVATE_KEY",
-    "GOOGLE_PRIVATE_KEY_ID",
-    "GOOGLE_PROJECT_ID",
-    "OPENROUTER_API_KEY",
-    "OVHCLOUD_API_KEY",
-    "PERPLEXITY_API_KEY",
-    "PORTKEY_GATEWAY_URL",
-] as const satisfies readonly (keyof CloudflareBindings)[];
-
-function syncTextEnvironment(env: CloudflareBindings): void {
-    // Text provider config still reads process.env. In Workers all bindings are
-    // stable per deployment, so copying known string bindings before generation
-    // is deterministic across concurrent requests in the same isolate.
-    for (const key of TEXT_ENV_KEYS) {
-        const value = env[key];
-        if (typeof value === "string") {
-            process.env[key] = value;
-        }
-    }
-}
 
 function generatePollinationsId(): string {
     return `pllns_${crypto.randomUUID().replaceAll("-", "")}`;
@@ -336,31 +292,6 @@ function base64ToArrayBuffer(value: string): ArrayBuffer {
     return bytes.buffer;
 }
 
-function serializeDetails(details: unknown): string | undefined {
-    if (details === undefined || details === null) return undefined;
-    return typeof details === "string" ? details : JSON.stringify(details);
-}
-
-function throwTextError(error: ServiceError): never {
-    const status =
-        typeof error.status === "number"
-            ? error.status
-            : typeof error.code === "number"
-              ? error.code
-              : 500;
-
-    throw new UpstreamError(status as ContentfulStatusCode, {
-        message: error.message || "Text generation failed",
-        // Propagate only — the code is decided at the throw site.
-        errorCode: error.errorCode,
-        requestUrl: error.requestUrl,
-        upstreamStatus: error.upstreamStatus,
-        responseBody: serializeDetails(error.details || error.response?.data),
-        upstreamHeaders: error.upstreamHeaders,
-        cause: error,
-    });
-}
-
 async function generateTextResponse(
     c: TextContext,
     requestData: RequestData,
@@ -475,129 +406,6 @@ export async function handleChatCompletionLocal(
     body: CreateChatCompletionRequest,
 ): Promise<Response> {
     return generateTextResponse(c, getChatRequestData(body), false);
-}
-
-type DirectResponsesCandidate = FallbackCandidate & {
-    responsesTarget: DirectResponsesTarget;
-    originalIndex: number;
-};
-
-function directResponsesCandidates(
-    c: TextContext,
-    request: CreateResponseRequest,
-): DirectResponsesCandidate[] {
-    const candidates = fallbackCandidates(c.var.model);
-    const primary = candidates[0];
-    if (!primary || primary.communityEndpoint || primary.entry?.agentConfig) {
-        throw responsesInvalidRequest(
-            `Model ${request.model} does not support the direct Responses API`,
-        );
-    }
-    const primaryTarget = resolveDirectResponsesTarget(
-        primary.id,
-        request,
-        c.env,
-    );
-    if (!primaryTarget) {
-        throw responsesInvalidRequest(
-            `Model ${request.model} does not support the direct Responses API`,
-        );
-    }
-
-    const supported: DirectResponsesCandidate[] = [
-        { ...primary, responsesTarget: primaryTarget, originalIndex: 0 },
-    ];
-    for (let index = 1; index < candidates.length; index += 1) {
-        const candidate = candidates[index];
-        if (candidate.communityEndpoint || candidate.entry?.agentConfig) {
-            continue;
-        }
-        const target = resolveDirectResponsesTarget(
-            candidate.id,
-            request,
-            c.env,
-        );
-        if (!target) continue;
-        supported.push({
-            ...candidate,
-            responsesTarget: target,
-            originalIndex: index,
-        });
-    }
-    return supported;
-}
-
-export async function handleCreateResponseLocal(
-    c: TextContext,
-    request: CreateResponseRequest,
-): Promise<Response> {
-    syncTextEnvironment(c.env);
-    validateDirectResponsesRequest(request);
-
-    try {
-        const { result, candidate } = await withModelFallback(
-            directResponsesCandidates(c, request),
-            (attempt) => callDirectResponses(request, attempt.responsesTarget),
-            c.var.track?.attempts,
-            (attempt) => enforceModelRateLimit(c, attempt),
-        );
-        c.set("upstreamRequestUrl", result.requestUrl);
-
-        const headers = new Headers({
-            "Content-Type": request.stream
-                ? "text/event-stream; charset=utf-8"
-                : "application/json; charset=utf-8",
-            "Cache-Control": request.stream ? "no-cache" : "no-store",
-            [MODEL_USED_HEADER]: candidate.id,
-        });
-        if (candidate.originalIndex > 0) {
-            headers.set(
-                FALLBACK_TARGET_HEADER,
-                formatFallbackTarget(candidate.originalIndex),
-            );
-        }
-
-        if (!request.stream) {
-            let data: unknown;
-            try {
-                data = await result.response.clone().json();
-            } catch (cause) {
-                throw new UpstreamError(502, {
-                    message: "Responses provider returned invalid JSON",
-                    requestUrl: result.requestUrl,
-                    cause,
-                });
-            }
-            if (
-                !data ||
-                typeof data !== "object" ||
-                (data as { object?: unknown }).object !== "response"
-            ) {
-                throw new UpstreamError(502, {
-                    message: "Responses provider returned an invalid response",
-                    requestUrl: result.requestUrl,
-                    responseBody: JSON.stringify(data),
-                });
-            }
-            const usage = getResponseUsage(data);
-            for (const [name, value] of Object.entries(
-                buildUsageHeaders(
-                    candidate.id,
-                    usage ? responsesUsageToUsage(usage) : undefined,
-                ),
-            )) {
-                headers.set(name, value);
-            }
-        }
-
-        const response = new Response(result.response.body, { headers });
-        if (!request.stream) {
-            c.var.track?.overrideResponseTracking(response.clone());
-        }
-        return response;
-    } catch (thrown) {
-        throwTextError(thrown as ServiceError);
-    }
 }
 
 export async function handleTextContentLocal(
