+++ gen.pollinations.ai/src/text/responses/client.ts
@@ -1,42 +1,12 @@
 import { collectUpstreamHeaders, remapUpstreamStatus } from "@shared/error.ts";
-import {
-    type CreateResponseRequest,
-    type ResponseUsage,
-    ResponseUsageSchema,
-} from "@shared/schemas/openai.ts";
-import { findModelByName } from "./availableModels.js";
-import type { ServiceError, TransformOptions } from "./types.js";
-import { resolveModelConfig } from "./utils/modelResolver.js";
+import type { CreateResponseRequest } from "@shared/schemas/openai.ts";
+import { findModelByName } from "../availableModels.js";
+import type { ServiceError, TransformOptions } from "../types.js";
+import { resolveModelConfig } from "../utils/modelResolver.js";
+import { buildDirectResponsesRequestBody } from "./request.js";
+import type { DirectResponsesTarget, JsonObject } from "./types.js";
 
 const REQUEST_TIMEOUT_MS = 290_000;
-
-type JsonObject = Record<string, unknown>;
-
-export type DirectResponsesTarget = {
-    authConfigured: boolean;
-    endpoint: string;
-    headers: Record<string, string>;
-    model: string;
-    defaults: JsonObject;
-};
-
-export function responsesInvalidRequest(
-    message: string,
-    param: string | null = "model",
-): ServiceError {
-    const error = new Error(message) as ServiceError;
-    error.status = 400;
-    error.errorCode = "invalid_request_error";
-    error.details = {
-        error: {
-            message,
-            type: "invalid_request_error",
-            code: "unsupported_parameter",
-            param,
-        },
-    };
-    return error;
-}
 
 function reasoningEffort(request: CreateResponseRequest): string | undefined {
     const effort = request.reasoning?.effort;
@@ -114,71 +84,6 @@ export function resolveDirectResponsesTarget(
     };
 }
 
-export function validateDirectResponsesRequest(
-    request: CreateResponseRequest,
-): void {
-    for (const [index, tool] of (request.tools ?? []).entries()) {
-        if (tool.type !== "function") {
-            throw responsesInvalidRequest(
-                "Only function tools are supported by the stateless direct Responses endpoint",
-                `tools[${index}].type`,
-            );
-        }
-    }
-
-    const pending: unknown[] = [request.input];
-    while (pending.length > 0) {
-        const value = pending.pop();
-        if (Array.isArray(value)) {
-            pending.push(...value);
-            continue;
-        }
-        if (!value || typeof value !== "object") continue;
-        const item = value as Record<string, unknown>;
-        if ("encrypted_content" in item || item.type === "item_reference") {
-            throw responsesInvalidRequest(
-                "Encrypted or reusable response state is not supported by the stateless Responses endpoint",
-                "input",
-            );
-        }
-        pending.push(...Object.values(item));
-    }
-}
-
-function requestBody(
-    request: CreateResponseRequest,
-    target: DirectResponsesTarget,
-): JsonObject {
-    const body: JsonObject = {
-        ...target.defaults,
-        ...request,
-        model: target.model,
-        store: false,
-    };
-    delete body.safe;
-
-    // SDKs often serialize inert state fields. Do not forward them because
-    // this endpoint has no Pollinations or provider-side response state.
-    if (body.previous_response_id == null) delete body.previous_response_id;
-    if (body.conversation == null) delete body.conversation;
-    if (body.background === false || body.background == null)
-        delete body.background;
-    if (Array.isArray(body.include) && body.include.length === 0)
-        delete body.include;
-    if (
-        Array.isArray(body.context_management) &&
-        body.context_management.length === 0
-    )
-        delete body.context_management;
-    if (body.prompt == null) delete body.prompt;
-    if (!body.stream) delete body.stream_options;
-
-    for (const [key, value] of Object.entries(body)) {
-        if (value === undefined || value === null) delete body[key];
-    }
-    return body;
-}
-
 function parseJson(text: string): unknown {
     try {
         return JSON.parse(text);
@@ -215,6 +120,7 @@ export async function callDirectResponses(
         error.requestUrl = requestUrl;
         throw error;
     }
+
     let response: Response;
     try {
         response = await fetcher(target.endpoint, {
@@ -223,7 +129,9 @@ export async function callDirectResponses(
                 "Content-Type": "application/json",
                 ...target.headers,
             },
-            body: JSON.stringify(requestBody(request, target)),
+            body: JSON.stringify(
+                buildDirectResponsesRequestBody(request, target),
+            ),
             signal: AbortSignal.timeout(REQUEST_TIMEOUT_MS),
         });
     } catch (thrown) {
@@ -258,11 +166,4 @@ export async function callDirectResponses(
         throw error;
     }
     return { response, requestUrl };
-}
-
-export function getResponseUsage(value: unknown): ResponseUsage | null {
-    if (!value || typeof value !== "object") return null;
-    const usage = (value as { usage?: unknown }).usage;
-    const parsed = ResponseUsageSchema.safeParse(usage);
-    return parsed.success ? parsed.data : null;
 }
