+++ gen.pollinations.ai/src/text/responses/types.ts
@@ -1,0 +1,9 @@
+export type JsonObject = Record<string, unknown>;
+
+export type DirectResponsesTarget = {
+    authConfigured: boolean;
+    endpoint: string;
+    headers: Record<string, string>;
+    model: string;
+    defaults: JsonObject;
+};
